requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
whitenoise==6.8.2
gunicorn==23.0.0
dj-database-url==2.3.0
python-dotenv==1.0.1
